requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
gspread==5.12.3
idna==3.6
oauthlib==3.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
urllib3==2.1.0
